removed converter util, added missing test case, added ModelMapper, add validation, added code formate

+++ src/main/java/com/example/demo/service/impl/UserContactServiceImpl.java
@@ -1,26 +1,27 @@
+/* (C)2023 */
 package com.example.demo.service.impl;
 
 import com.example.demo.data.UserContact;
 import com.example.demo.dto.UserContactRequest;
-import com.example.demo.dto.UserContactResponse;
 import com.example.demo.repository.UserContactRepository;
 import com.example.demo.service.UserContactService;
-import com.example.demo.uitls.UserContactUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.UUID;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 @Service
 public class UserContactServiceImpl implements UserContactService {
 
-    @Autowired
-    private UserContactRepository userContactRepository;
+    @Autowired private UserContactRepository userContactRepository;
+
+    @Autowired ModelMapper modelMapper;
 
     @Override
     public boolean addUserContactDetails(UserContactRequest userContactRequest) {
-        userContactRepository.save(UserContactUtils.CONVERT_USER_CONTACT_REQUEST_TO_ENTITY.apply(userContactRequest));
+        UserContact userContact = modelMapper.map(userContactRequest, UserContact.class);
+        userContactRepository.save(userContact);
         return true;
     }
 
@@ -31,7 +32,9 @@ public class UserContactServiceImpl implements UserContactService {
 
     @Override
     public boolean deleteUserDetailsById(UUID id) {
-        userContactRepository.deleteById(id);
-        return true;
+        if (userContactRepository.findById(id).isPresent()) {
+            userContactRepository.deleteById(id);
+            return true;
+        } else return false;
     }
 }
